refactor: extract AddAccountDrawer into its own folder
