+++ apps/ledger-live-desktop/src/newArch/features/AddAccountDrawer/ModularDrawerAddAccountFlowManager.tsx
@@ -8,18 +8,18 @@ import { useTranslation } from "react-i18next";
 import styled from "styled-components";
 import { ADD_ACCOUNT_FLOW_NAME } from "./analytics/addAccount.types";
 import { MODULAR_DRAWER_PAGE_NAME } from "./analytics/modularDrawer.types";
-import AnimatedScreenWrapper from "./components/AnimatedScreenWrapper";
-import { BackButtonArrow } from "./components/BackButton";
-import HeaderGradient from "./components/HeaderGradient";
-import { useAddAccountFlowNavigation } from "./hooks/useAddAccountFlowNavigation";
+import AnimatedScreenWrapper from "../ModularDrawer/components/AnimatedScreenWrapper";
+import { BackButtonArrow } from "../ModularDrawer/components/BackButton";
+import HeaderGradient from "../ModularDrawer/components/HeaderGradient";
+import { useAddAccountFlowNavigation } from "./useAddAccountFlowNavigation";
 import AccountsAdded from "./screens/AccountsAdded";
-import { AccountSelection } from "./screens/AccountSelection";
+import { AccountSelection } from "../ModularDrawer/screens/AccountSelection";
 import AccountsWarning from "./screens/AccountsWarning";
 import ConnectYourDevice from "./screens/ConnectYourDevice";
 import EditAccountName from "./screens/EditAccountName";
 import FundAccount from "./screens/FundAccount";
 import ScanAccounts from "./screens/ScanAccounts";
-import { MODULAR_DRAWER_ADD_ACCOUNT_STEP, ModularDrawerAddAccountStep } from "./types";
+import { MODULAR_DRAWER_ADD_ACCOUNT_STEP, ModularDrawerAddAccountStep } from "./domain";
 
 const ANALYTICS_PROPERTY_FLOW = "Modular Add Account Flow";
