+++ apps/ledger-live-desktop/src/newArch/features/AddAccountDrawer/analytics/TrackAddAccountScreen.tsx
