+++ apps/ledger-live-desktop/src/newArch/features/AddAccountDrawer/domain.ts
@@ -1,0 +1,21 @@
+export const MODULAR_DRAWER_ADD_ACCOUNT_STEP = {
+  CONNECT_YOUR_DEVICE: "CONNECT_YOUR_DEVICE",
+  SCAN_ACCOUNTS: "SCAN_ACCOUNTS",
+  ACCOUNTS_ADDED: "ACCOUNTS_ADDED",
+  EDIT_ACCOUNT_NAME: "EDIT_ACCOUNT_NAME",
+  ACCOUNTS_WARNING: "ACCOUNTS_WARNING",
+  FUND_ACCOUNT: "FUND_ACCOUNT",
+  SELECT_ACCOUNT: "SELECT_ACCOUNT",
+} as const;
+
+export type ModularDrawerAddAccountStep =
+  (typeof MODULAR_DRAWER_ADD_ACCOUNT_STEP)[keyof typeof MODULAR_DRAWER_ADD_ACCOUNT_STEP];
+
+export const MODULAR_DRAWER_ADD_ACCOUNT_CATEGORY = "AddAccounts";
+
+export const WARNING_REASON = {
+  ALREADY_EMPTY_ACCOUNT: "ALREADY_EMPTY_ACCOUNT",
+  NO_ASSOCIATED_ACCOUNTS: "NO_ASSOCIATED_ACCOUNTS",
+} as const;
+
+export type WarningReason = (typeof WARNING_REASON)[keyof typeof WARNING_REASON];
