+++ apps/ledger-live-desktop/src/newArch/features/AddAccountDrawer/screens/AccountsAdded/components/SuccessIcon.tsx
