+++ apps/ledger-live-desktop/src/newArch/features/AddAccountDrawer/screens/AccountsWarning/components/index.ts
@@ -1,2 +1,2 @@
-export { IconContainer } from "./IconContainer";
 export { ActionButtons } from "./ActionButtons";
+export { IconContainer } from "./IconContainer";
