+++ apps/ledger-live-desktop/src/newArch/features/AddAccountDrawer/screens/ConnectYourDevice/index.tsx
@@ -1,12 +1,12 @@
 import type { AppResult } from "@ledgerhq/live-common/hw/actions/app";
+import { CryptoCurrency } from "@ledgerhq/types-cryptoassets";
 import React, { useEffect } from "react";
+import { useSelector } from "react-redux";
+import { useOnDemandCurrencyCountervalues } from "~/renderer/actions/deprecated/ondemand-countervalues";
 import { HOOKS_TRACKING_LOCATIONS } from "~/renderer/analytics/hooks/variables";
 import { prepareCurrency } from "~/renderer/bridge/cache";
 import DeviceAction from "~/renderer/components/DeviceAction";
 import useConnectAppAction from "~/renderer/hooks/useConnectAppAction";
-import { CryptoCurrency } from "@ledgerhq/types-cryptoassets";
-import { useOnDemandCurrencyCountervalues } from "~/renderer/actions/deprecated/ondemand-countervalues";
-import { useSelector } from "react-redux";
 import { counterValueCurrencySelector } from "~/renderer/reducers/settings";
 import { TrackAddAccountScreen } from "../../analytics/TrackAddAccountScreen";
 import { ADD_ACCOUNT_FLOW_NAME, ADD_ACCOUNT_PAGE_NAME } from "../../analytics/addAccount.types";
