+++ apps/ledger-live-desktop/src/newArch/features/AddAccountDrawer/screens/FundAccount/index.tsx
@@ -1,12 +1,12 @@
-import React from "react";
-import { Account } from "@ledgerhq/types-live";
-import { useAccountFormatter } from "../AccountsAdded/hooks";
+import { Flex } from "@ledgerhq/react-ui/index";
 import { AccountItem } from "@ledgerhq/react-ui/pre-ldls/index";
-import { ActionsContainer } from "./components";
 import { CryptoCurrency } from "@ledgerhq/types-cryptoassets";
+import { Account } from "@ledgerhq/types-live";
+import React from "react";
 import { TrackAddAccountScreen } from "../../analytics/TrackAddAccountScreen";
 import { ADD_ACCOUNT_FLOW_NAME, ADD_ACCOUNT_PAGE_NAME } from "../../analytics/addAccount.types";
-import { Flex } from "@ledgerhq/react-ui/index";
+import { useAccountFormatter } from "../AccountsAdded/hooks";
+import { ActionsContainer } from "./components";
 
 interface Props {
   account: Account;
