+++ apps/ledger-live-desktop/src/newArch/features/AddAccountDrawer/screens/ScanAccounts/useScanAccounts.ts
