+++ apps/ledger-live-desktop/src/renderer/screens/accounts/AccountsHeader.tsx
@@ -9,8 +9,8 @@ import { EntryPoint } from "LLD/features/LedgerSyncEntryPoints/types";
 import { useOpenAssetFlow } from "LLD/features/ModularDrawer/hooks/useOpenAssetFlow";
 import { ModularDrawerLocation } from "LLD/features/ModularDrawer";
 import { MAD_SOURCE_PAGES } from "LLD/features/ModularDrawer/analytics/modularDrawer.types";
-import useAddAccountAnalytics from "LLD/features/ModularDrawer/analytics/useAddAccountAnalytics";
-import { ADD_ACCOUNT_EVENTS_NAME } from "LLD/features/ModularDrawer/analytics/addAccount.types";
+import useAddAccountAnalytics from "LLD/features/AddAccountDrawer/analytics/useAddAccountAnalytics";
+import { ADD_ACCOUNT_EVENTS_NAME } from "LLD/features/AddAccountDrawer/analytics/addAccount.types";
 
 const AccountsHeader = () => {
   const { t } = useTranslation();
